refactor: 당첨 번호 관련 기능을 app에서 WinningLottoManager로 분리
app이 여러 역할을 수행한다고 판단하여, WinningLottoManager라는 클래스를 생성해 역할 분리
